@@ -1,11 +1,11 @@
 module codegen
 
-go 1.23
+go 1.24
 
 require (
 	github.com/flosch/pongo2/v6 v6.0.0
 	github.com/iancoleman/strcase v0.3.0
-	github.com/tidwall/gjson v1.17.3
+	github.com/tidwall/gjson v1.18.0
 )
 
 require (
